방명록 모델 및 컨트롤 수정

+++ guestbook/views.py
@@ -7,17 +7,18 @@ from guestbook.models import Guestbook
 
 def index(request):
     guestbook_list = Guestbook.objects.all().order_by('-regdate')
-
+    print(guestbook_list)
     context = {'guestbook_list' : guestbook_list}
-    return render(request, 'guestbook/index.html', context)
+
+    return render(request, 'guestbook/list.html', context)
 
 def add (request):
     guestbook = Guestbook()
-    guestbook.name = request.POST['name']
-    guestbook.password = request.POST['password']
-    guestbook.message = request.POST['message']
-
-    guestbook.save()
+    if request.POST['a'] == 'insert':
+        guestbook.name = request.POST['name']
+        guestbook.password = request.POST['pass']
+        guestbook.message = request.POST['content']
+        guestbook.save()
 
     return HttpResponseRedirect('/guestbook')
 
@@ -27,11 +28,9 @@ def deleteform(request):
     return render(request, 'guestbook/deleteform.html', id)
 
 def delete(request):
-    id = request.POST['id']
-    password = request.POST['password']
-    # try:
-    context = Guestbook.objects.filter(id=id).filter(password=password).delete()
-    # except ValueError as vE:
-    #     print('[Error name : {0}, Occured time : {1}'.format(vE, datetime.now()))
+    if request.POST['a'] == 'delete':
+        id = request.POST['id']
+        password = request.POST['password']
+        Guestbook.objects.filter(id=id).filter(password=password).delete()
 
-    return HttpResponseRedirect('/guestbook')
+        return HttpResponseRedirect('/guestbook')
